+++ test/utils.tsx
@@ -11,6 +11,7 @@ import { expect, vi } from 'vitest';
 
 import configureHooks from '../src/hooks/index.js';
 import configureQueryClient from '../src/queryClient.js';
+import { DEBOUNCE_TIME_AUTORESIZE } from '../src/config/constants.js';
 import { Notifier, QueryClientConfig } from '../src/types.js';
 import { API_HOST, MOCK_APP_ORIGIN, RequestMethods, WS_HOST } from './constants.js';
 
@@ -36,6 +37,7 @@ export const setUpTest = (args?: Args) => {
     WS_HOST,
     enableWebsocket: false,
     isStandalone: false,
+    debounceTimeAutoResize: DEBOUNCE_TIME_AUTORESIZE,
   };
 
   const { queryClient, QueryClientProvider, mutations, useMutation } =